Allow upgrade info to be defined by env variables
When bootstrapping a node from scratch, after an upgrade, the built-in
cosmovisor setup in the container doesn't have any means to detect
what's the correct version to start with.

This commit declares two environment variables, `$UPGRADE_VERSION` and
`$UPGRADE_HEIGHT` which can be declared for state-sync bootstrap.

The node will then start with the right node binary version and continue
as if it was running since genesis.