Embed cosmovisor upgrade manager in container (#20)
In order to manage on-chain upgrades, build cosmovisor into the container
image together with all node binaries so that single container can run
through the whole chain from genesis to the HEAD.

When bootstrapping a node from scratch, after an upgrade, using state-sync,
the built-in cosmovisor setup in the container doesn't have any means to
detect what's the correct version to start with.

This change declares two environment variables, `$UPGRADE_VERSION` and
`$UPGRADE_HEIGHT` which can be declared for state-sync bootstrap.

The node will then start with the right node binary version and continue
as if it was running since genesis.